Just include a file directly.
